# floppa-web-shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Floppa VPN Admin API",
    "description": "API for Floppa VPN admin panel and user management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/account/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log in with a login + password.",
        "operationId": "login_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid login or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/account/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a new account with a login + password (no Telegram). Grants a short taster trial.",
        "operationId": "register_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid login or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Login already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/telegram": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate via Telegram Login Widget",
        "operationId": "telegram_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramAuthData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid Telegram auth data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/telegram/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Telegram widget callback for deep-link flow.\nReturns an HTML landing page that auto-opens the app via deep link,\nwith a manual button and copy-code fallback for browsers that block custom schemes.",
        "operationId": "telegram_deep_link_callback",
        "responses": {
          "200": {
            "description": "HTML page that redirects to deep link"
          },
          "400": {
            "description": "Invalid or expired state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Invalid Telegram auth payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/telegram/exchange-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange one-time login code for JWT + user payload.",
        "operationId": "exchange_telegram_login_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeTelegramLoginCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/telegram/mini-app": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate via Telegram Mini App initData",
        "operationId": "telegram_mini_app_auth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid Mini App initData",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/telegram/start": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Render the Telegram login page for deep-link flow.",
        "operationId": "start_telegram_deep_link_login",
        "parameters": [
          {
            "name": "redirect_uri",
            "in": "query",
            "description": "Deep link URI, e.g. floppa://auth",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML login page"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Server misconfiguration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountLoginRequest": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AccountRegisterRequest": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserInfo"
          }
        }
      },
      "AuthUserInfo": {
        "type": "object",
        "required": [
          "id",
          "is_admin"
        ],
        "properties": {
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "is_admin": {
            "type": "boolean"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "photo_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ExchangeTelegramLoginCodeRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
...
      "MiniAppAuthRequest": {
        "type": "object",
        "required": [
          "init_data"
        ],
        "properties": {
          "init_data": {
            "type": "string"
          }
        }
      },
...
      "TelegramAuthData": {
        "type": "object",
        "description": "Data received from Telegram Login Widget",
        "required": [
          "id",
          "auth_date",
          "hash"
        ],
        "properties": {
          "auth_date": {
            "type": "integer",
            "format": "int64"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "hash": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "photo_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}